Update bot.py
Changed the print statements to use the logger package.
I also added a function which would automatically create a array of strings containing the syntax that the load_extension method likes for all the files in the cog folder which end in "Cog.py". So if it is not actually a cog then an error will be thrown.